General Ideas
The solution would be:
Create a service that to do the exact API call, and based on the
response, using RXJS formating methods like 'switchMap', etc to return
the response in the array format.
Then inject this service into the NGRX effect, the effect would based
on the action type, assign the returned response back to state. And
then the profile component can subscribe to and update the component
accordingly.

To make a new page, the pattern would be the same with the solution
above, except pass numbers of user generated, and then save it to the
store.  When clicked, first to reroute to the detail page, then,
mapping the user to the 'activeUser' object.
Use the same service to fetch the user data, if not available.